+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.0] - 2022-09-06
+
+### Added
+- VersionRange::isEqualTo()
+- VersionRange::isHigherThan()
+- VersionRange::isLowerThan()
+
 ## [3.3.0] - 2022-06-18
 
 ### Added